requirements.txt
Django==1.6.5
South==0.8.4
argparse==1.2.1
djangorestframework==2.3.13
wsgiref==0.1.2
